refactor: moved all form logics and states
